+++ CHANGELOG.md
@@ -178,3 +178,4 @@
 - Show act paths in the apk view and can replay act paths
 - Rewrite controllers to decouple web_router
 - Add act paths to record testcase
+- Add test cases to record testcase